% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggobs.R
\name{aggregate_observations}
\alias{aggregate_observations}
\title{Aggregate observations hierarchically with False Split Rate control}
\usage{
aggregate_observations(
  y,
  sigma_constant = NULL,
  hc = NULL,
  dend = NULL,
  hc_list = NULL,
  alpha_level
)
}
\arguments{
\item{y}{A length-\code{n-observation} vector of observations}

\item{sigma_constant}{Standard deviation of noise.}

\item{hc}{An object of class \code{hclust}.}

\item{dend}{An object of class \code{dendrogram}.}

\item{hc_list}{A list of length-\code{num_interior_nodes} where the ith item in the list contains the child nodes of the ith node in the tree. The negative values in the list indicate leaf nodes. When \code{hc_list} is \code{NULL}, function will learn it from \code{hc} or \code{dend}.}

\item{alpha_level}{A use-specified target FSR level}
}
\value{
Returns the aggregation result.
\item{alpha}{The target FSR level.}
\item{groups}{A length-\code{n-observation} vector of integers indicating the cluster to which each observation is allocated.}
\item{rejections}{A length-(\code{num_interior_nodes}) vector indicating whether each node is rejected.}
}
\description{
This function aggregate observations with the same means while simultaneously controlling False Split Rate under a target level.
The aggregation is achieved by two steps: (1) Generate p-values for each interior node through ANOVA test (2) Sequentially test on the tree.
}
\examples{
## See data example.
}
