% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsp_and_power.R
\name{calculate_power}
\alias{calculate_power}
\title{Calculate Power}
\usage{
calculate_power(theta_est, theta_true)
}
\arguments{
\item{theta_est}{A vector of integers indicating the cluster to which each item is allocated by an achieved aggregation.}

\item{theta_true}{A vector of integers indicating the cluster to which each item is allocated by the true aggregation.}
}
\value{
Power.
}
\description{
This function calculate power of an achieved aggregation given the true aggregation.
}
\examples{
v1 = c(1,1,2,2,2,3,3,3,3,4,4,4)
v2 = c(1,1,1,1,1,2,2,2,3,4,4,4)
calculate_power(v1, v2)
## 0.6666667
}
