% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dend_as_hclist.R
\name{dend_as_hclist}
\alias{dend_as_hclist}
\title{Transform a dendrogram into a list object}
\usage{
dend_as_hclist(dend)
}
\arguments{
\item{dend}{An object of class \code{dendrogram}.}
}
\value{
A list object of length-\code{num_interior_nodes}. The \code{i}-th item in the list contains the child nodes of the \code{i}-th node in the tree. The negative values in the list indicate leaf nodes, and positive values indicate interior node. The interior nodes are numbered reversely from the root along the branch, i.e., the number corresponding to a node is always smaller than the number corresponding to its parent.
\item{hc_list}{The list object that represents the tree structure,}
\item{leaf_labels}{A length-\code{n-leaf} of labels of leaves stored in the \code{dendrogram}.}
}
\description{
This function transforms a dendrogram into a hc_list object
}
\examples{
hc <- hclust(dist(USArrests), "ave")
dend = as.dendrogram(hc)
dah = dend_as_hclist(dend)
## check length of the list
length(dah$hc_list)
## leaf labels
dah$leaf_labels
}
