% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclist_as_dend.R
\name{hclist_as_dend}
\alias{hclist_as_dend}
\title{Transform a hc_list object into a dendrogram}
\usage{
hclist_as_dend(hc_list)
}
\arguments{
\item{hc_list}{An object of class \code{hc_list}.}
}
\value{
An object of class \code{dendrogram} that stores a tree structure.
\item{dend}{An object of class \code{dendrogram}.}
}
\description{
This function transforms a dendrogram into a hc_list object
}
\examples{
hc <- hclust(dist(USArrests), "ave")
dend = as.dendrogram(hc)
hclist = dend_as_hclist(dend)
dend2 = hclist_as_dend(hc_list = hclist)
plot(dend2, center = TRUE)
# can compare with plot(dend)
}
