% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_test.R
\name{hierarchical_test}
\alias{hierarchical_test}
\title{Test on a given tree to achieve aggregation of leaves}
\usage{
hierarchical_test(
  hc = NULL,
  dend = NULL,
  hc_list = NULL,
  p_vals,
  alpha_level,
  independent = TRUE
)
}
\arguments{
\item{hc}{An object of class \code{hclust}.}

\item{dend}{An object of class \code{dendrogram}.}

\item{hc_list}{A list of length-\code{num_interior_nodes} where the ith item in the list contains the child nodes of the ith node in the tree. The negative values in the list indicate leaf nodes. When \code{hc_list} is \code{NULL}, function will learn it from \code{hc} or \code{dend}.}

\item{p_vals}{A length-\code{num_interior_nodes} vector of p-values corresponding to interior nodes.}

\item{alpha_level}{A use-specified target FSR level}

\item{independent}{Whether the p-values are independent (default = TRUE).}
}
\value{
Returns the testing result and calculated threshold values.
\item{alpha}{The target FSR level.}
\item{rejections}{A length-\code{num_interior_nodes} vector indicating whether each node is rejected.}
\item{threshold_functions}{A length-(\code{nnodes}-\code{nleaves}) vector of computed threshold value at each node. NA if not tested.}
\item{groups}{A length-\code{n-feature} vector of integers indicating the cluster to which each leaf is allocated.}
}
\description{
This function sequentially tests on a tree in a top-down manner. The testing result determines aggregation of the leaves and make sure the False Split Rate is controlled under a target level.
}
\examples{
set.seed(123)
hc = hclust(dist((1:10) + runif(10)/10), method = "complete")
p_vals = c(runif(5), rbeta(4, 1, 60))
alpha_level = 0.3
independent = TRUE
hierarchical_test(hc = hc, dend = NULL, p_vals = p_vals, alpha_level = alpha_level, independent = independent)
}
