% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_hc_list}
\alias{plot_hc_list}
\title{Visualize the Tree Structure}
\usage{
plot_hc_list(
  hc_list = NULL,
  iflabel = FALSE,
  labels = NULL,
  hc = NULL,
  dend = NULL
)
}
\arguments{
\item{hc_list}{A list of length-\code{num_interior_nodes} where the ith item in the list contains the child nodes of the ith node in the tree. The negative values in the list indicate leaf nodes. When \code{hc_list} is \code{NULL}, function will learn it from \code{hc} or \code{dend}.}

\item{iflabel}{A boolean variable indicating whether leaves should be labeled.}

\item{labels}{A vector of length-\code{n-leaf} of labels corresponding to each leaf.}

\item{hc}{An object of class \code{hclust}. Only required if \code{dend} is NULL.}

\item{dend}{An object of class \code{dendrogram}. When dend is NULL, the function will learn \code{dend} from \code{hc}.}
}
\description{
This function plots a tree structure. The tree structure can be represented by a variety of object formats: \code{dendrogram}, \code{hclust} or a list (\code{hc_list}).
}
\examples{
## Example 1: Plot an hc object
hc = hclust(dist(USArrests), "ave")
labels = rownames(USArrests)[hc$order]
plot_hc_list(hc = hc, iflabel = TRUE)

## Example 2: Plot a dendrogram object
dend = as.dendrogram(hc)
plot_hc_list(dend = dend, iflabel = TRUE)
## Example 3: Plot a hc_list object
dah = dend_as_hclist(dend)
plot_hc_list(hc_list = dah$hc_list, iflabel = TRUE, labels = dah$leaf_labels)
}
