% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stocks_tree}
\alias{stocks_tree}
\title{Hierarchical tree based on NAICS for companies in the stocks dataset}
\format{
An object of class \code{list} of length 702.
}
\source{
Information of North American Industry Classification System available at \url{https://www.census.gov/naics/?58967?yearbck=2017}. Company's NAICS code from CRSP/Compustat merged database.
}
\usage{
stocks_tree
}
\description{
A list of length-\code{702} that stores the tree structure (referred to as \code{hc_list} format in the package) for the 2538 stocks. The ith item in the list contains the child nodes of the ith node in the tree. The negative values in the list indicate leaf nodes and the positive values correspond to interior nodes.
The tree structure is formed based on North American Industry Classification System (NAICS).
}
\keyword{datasets}
