% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stocks_volatility}
\alias{stocks_volatility}
\title{Stocks volatility}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2538 rows and 3 columns.
}
\source{
Center for Research in Security Prices (CRSP)
}
\usage{
stocks_volatility
}
\description{
A dataframe of 2538 stocks' volatility, as well as their corresponding company name. The volatility is estimated by daily high and low price and averaged throughout 2015-2019 (See paper "Controlling the False Split Rate in Tree-Based Aggregation" for more details).
}
\keyword{datasets}
