% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxi_tree}
\alias{taxi_tree}
\title{Hierarchical tree based on neighborhoods in NYC}
\format{
An object of class \code{list} of length 190.
}
\usage{
taxi_tree
}
\description{
A list of length-\code{190} that stores the tree structure (referred to as \code{hc_list} format in the package) for the 194 neighborhoods in New York city. The ith item in the list contains the child nodes of the ith node in the tree. The negative values in the list indicate leaf nodes and the positive values correspond to interior nodes.
The tree structure is formed based on Neighborhood Tabulation Areas (NTAs).
}
\keyword{datasets}
