% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggfeature.R
\name{aggregate_features}
\alias{aggregate_features}
\title{Aggregate rare features hierarchically with False Split Rate control}
\usage{
aggregate_features(y, X, sigma = NULL, tree = NULL, alpha)
}
\arguments{
\item{y}{A length-\code{n-observation} response vector that guides the aggregation}

\item{X}{An \code{n-observation}-by-\code{n-feature} design matrix. Each row corresponds to a subject and each column stores the observation of a feature made by subjects.}

\item{sigma}{Standard deviation of noise. If not given, the algorithm will estimate sigma.}

\item{tree}{An object encoding the tree structure.  Can be one of three
formats: (1) an \code{hclust} object (if tree is binary), (2) a \code{dendrogram}, or
(3) a generalization of an hclust object to the case of non-binary trees,
which we call an hc_list object. An hc_list object is a list of
length-\code{num_interior_nodes} where the ith item in the list contains
the child nodes of the ith node in the tree. The negative values in the
list indicate leaf nodes.}

\item{alpha}{A use-specified target FSR level}
}
\value{
Returns the aggregation result.
\item{alpha}{The target FSR level.}
\item{groups}{A length-\code{n-feature} vector of integers indicating the cluster to which each feature is allocated.}
\item{rejections}{A length-(\code{num_interior_nodes}) vector indicating whether each node is rejected.}
\item{p_vals}{A length-(\code{num_interior_nodes}) vector of the p-value (note: all are computed although not all are used in the sequential testing procedure)}
}
\description{
This function achieves rare features aggregation while simultaneously controlling False Split Rate under a target level.
The aggregation is achieved by two steps: (1) Generate p-values for each interior node (2) Sequentially test on the tree.
}
\examples{
## See vignette for a small data example.
}
