% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggfeature.R
\name{aggregate_features}
\alias{aggregate_features}
\title{Aggregate rare features hierarchically with False Split Rate control}
\usage{
aggregate_features(
  y,
  X,
  sigma_constant = NULL,
  hc = NULL,
  dend = NULL,
  hc_list = NULL,
  alpha_level
)
}
\arguments{
\item{y}{A length-\code{n-observation} response vector that guides the aggregation}

\item{X}{An \code{n-observation}-by-\code{n-feature} design matrix. Each row corresponds to a subject and each column stores the observation of a feature made by subjects.}

\item{sigma_constant}{Standard deviation of noise.}

\item{hc}{An object of class \code{hclust}.}

\item{dend}{An object of class \code{dendrogram}.}

\item{hc_list}{A list of length-\code{num_interior_nodes} where the ith item in the list contains the child nodes of the ith node in the tree. The negative values in the list indicate leaf nodes. When \code{hc_list} is \code{NULL}, function will learn it from \code{hc} or \code{dend}.}

\item{alpha_level}{A use-specified target FSR level}
}
\value{
Returns the aggregation result.
\item{alpha}{The target FSR level.}
\item{groups}{A length-\code{n-feature} vector of integers indicating the cluster to which each feature is allocated.}
\item{rejections}{A length-(\code{num_interior_nodes}) vector indicating whether each node is rejected.}
}
\description{
This function achieves rare features aggregation while simultaneously controlling False Split Rate under a target level.
The aggregation is achieved by two steps: (1) Generate p-values for each interior node (2) Sequentially test on the tree.
}
\examples{
## See vignette for data example.
}
