% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_leaves.R
\name{find_leaves_in_list}
\alias{find_leaves_in_list}
\title{Find the leaves under a node}
\usage{
find_leaves_in_list(ind, hc_list)
}
\arguments{
\item{ind}{Index of a tree node.}

\item{hc_list}{A list of length-\code{num_interior_nodes} where the ith item in the list contains the child nodes of the ith node in the tree. The negative values in the list indicate leaf nodes.}
}
\value{
Returns the leaf indices.
}
\description{
This function finds all leaves that are descendants of a node in a tree (represented by an \code{hc_list} object).
}
\examples{
## See data example.
}
