% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_aggregation}
\alias{plot_aggregation}
\title{Visualize Aggregation on the Tree}
\usage{
plot_aggregation(
  rejections,
  iflabel = FALSE,
  labels = NULL,
  hc_list = NULL,
  hc = NULL,
  dend = NULL
)
}
\arguments{
\item{rejections}{A vector of length-\code{num_interior_nodes} indicating if each node is rejected. When coloring with true aggregation, this vector indicate whether each interior node is a non-null node.}

\item{iflabel}{A boolean variable indicating whether leaves should be labeled.}

\item{labels}{A vector of length-\code{n-leaf} of labels corresponding to each leaf.}

\item{hc_list}{A list of length-\code{num_interior_nodes} where the ith item in the list contains the child nodes of the ith node in the tree. The negative values in the list indicate leaf nodes. When \code{hc_list} is \code{NULL}, function will learn it from \code{hc} or \code{dend}.}

\item{hc}{An object of class \code{hclust}. Only required if \code{dend} is NULL.}

\item{dend}{An object of class \code{dendrogram}. When dend is NULL, the function will learn \code{dend} from \code{hc}.
\subsection{Example 1: Plot an hc object}{

hc = hclust(dist(USArrests), "ave")
labels = rownames(USArrests)\link{hc$order}
plot_aggregation(rejections = c(rep(FALSE, 45), iflabel = TRUE, rep(TRUE, 4)), hc = hc)
}

\subsection{Example 2: Plot a dendrogram object}{

hc = hclust(dist(USArrests), "ave")
dend = as.dendrogram(hc)
plot_aggregation(rejections = c(rep(FALSE, 45), rep(TRUE, 4)), iflabel = TRUE, dend = dend)
}

\subsection{Example 3: Plot a hc_list object}{

hc = hclust(dist(USArrests), "ave")
dend = as.dendrogram(hc)
dah = dend_as_hclist(dend)
plot_aggregation(rejections = c(rep(FALSE, 45), rep(TRUE, 4)), iflabel = TRUE, labels = dah$leaf_labels, hc_list = dah$hc_list)
}}
}
\description{
This function plots a tree structure with branches and leaves colored by group memberships. The tree structure can be represented by a variety of object formats: \code{dendrogram}, \code{hclust} or a a list (\code{hc_list}). The aggregation is determined by a vector corresponding to each interior nodes
}
