% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_tree}
\alias{plot_tree}
\title{Visualize the Tree Structure}
\usage{
plot_tree(tree = NULL, iflabel = FALSE, labels = NULL)
}
\arguments{
\item{tree}{An hclust (if binary tree), or dendrogram, or hc_list object that
stores the tree structure. An hc_list object is a list of
length-\code{num_interior_nodes} where the ith item in the list contains
the child nodes of the ith node in the tree. The negative values in the
list indicate leaf nodes.}

\item{iflabel}{A boolean variable indicating whether leaves should be labeled.}

\item{labels}{A vector of length-\code{n-leaf} of labels corresponding to each leaf.}
}
\description{
This function plots a tree structure. The tree structure can be represented by a variety of object formats: \code{dendrogram}, \code{hclust} or a list (\code{hc_list}).
}
\examples{
## Example 1: Plot an hc object
hc = hclust(dist(USArrests), "ave")
labels = rownames(USArrests)[hc$order]
plot_tree(tree = hc, iflabel = TRUE)
## Example 2: Plot a dendrogram object
dend = as.dendrogram(hc)
plot_tree(tree = dend, iflabel = TRUE)
## Example 3: Plot a hc_list object
dah = dend_as_hclist(dend)
labels = attr(dah, "leaf_labels")
plot_tree(tree = dah, iflabel = TRUE, labels = labels)
}
